separate cli and validation function files
