+++ pyproject.toml
@@ -49,7 +49,7 @@ dynamic = ["version"] # the build backend will compute the version dynamically f
 "Bug Tracker" = "https://github.com/NeurodataWithoutBorders/pynwb/issues"
 
 [project.scripts]
-pynwb-validate = "pynwb.validation:validate_cli"
+pynwb-validate = "pynwb.validation_cli:validation_cli"
 
 [tool.hatch.version]
 source = "vcs"
